# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted firmware registry for ESP32 OTA updates with webhook support",
        "title": "Firmware Registry API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://github.com/your-org/firmware-registry",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/firmware/{type}": {
            "get": {
                "security": [
                    {
                        "DeviceKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all firmware versions for a specific type, sorted by semantic version (newest first)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "firmware"
                ],
                "summary": "List firmware versions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Firmware type (e.g., esp32-main)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/firmware-registry-api_internal_firmware.FirmwareDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/firmware/{type}/latest": {
            "get": {
                "security": [
                    {
                        "DeviceKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the latest firmware version for a specific type based on semantic versioning",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "firmware"
                ],
                "summary": "Get latest firmware",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Firmware type (e.g., esp32-main)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firmware-registry-api_internal_firmware.FirmwareDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "No firmware found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/firmware/{type}/{version}": {
            "get": {
                "security": [
                    {
                        "DeviceKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download the firmware binary for a specific type and version",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "firmware"
                ],
                "summary": "Download firmware",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Firmware type (e.g., esp32-main)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Semantic version (e.g., 1.2.3)",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Firmware binary file",
                        "schema": {
                            "type": "file"
                        },
                        "headers": {
                            "X-Firmware-Sha256": {
                                "type": "string",
                                "description": "SHA256 checksum of the firmware"
                            },
                            "X-Firmware-Version": {
                                "type": "string",
                                "description": "Firmware version"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Firmware not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a new firmware binary for a specific type and version",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "firmware"
                ],
                "summary": "Upload firmware",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Firmware type (e.g., esp32-main)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Semantic version (e.g., 1.2.3)",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Firmware binary file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firmware-registry-api_internal_firmware.FirmwareDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid multipart or missing file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Save failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a firmware binary and its metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "firmware"
                ],
                "summary": "Delete firmware",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Firmware type (e.g., esp32-main)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Semantic version (e.g., 1.2.3)",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletion confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Firmware not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/webhooks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all registered webhooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "List webhooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/firmware-registry-api_internal_webhook.WebhookDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new webhook endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Create webhook",
                "parameters": [
                    {
                        "description": "Webhook configuration",
                        "name": "webhook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/firmware-registry-api_internal_webhook.WebhookDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created webhook ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or missing required fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/webhooks/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing webhook configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Update webhook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated webhook configuration",
                        "name": "webhook",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/firmware-registry-api_internal_webhook.WebhookDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or webhook ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a webhook from the registry",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Delete webhook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletion confirmation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid webhook ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "firmware-registry-api_internal_firmware.FirmwareDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "downloadUrl": {
                    "type": "string",
                    "example": "http://localhost:8080/api/firmware/esp32-main/1.2.3"
                },
                "filename": {
                    "type": "string",
                    "example": "firmware.bin"
                },
                "sha256": {
                    "type": "string",
                    "example": "abc123..."
                },
                "sizeBytes": {
                    "type": "integer",
                    "example": 524288
                },
                "type": {
                    "type": "string",
                    "example": "esp32-main"
                },
                "version": {
                    "type": "string",
                    "example": "1.2.3"
                }
            }
        },
        "firmware-registry-api_internal_webhook.WebhookDTO": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "firmware.uploaded",
                        "firmware.deleted"
                    ]
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/webhook"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Admin API key for administrative operations",
            "type": "apiKey",
            "name": "X-Admin-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT Bearer token from OIDC provider (format: \"Bearer {token}\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "DeviceKeyAuth": {
            "description": "Device API key for firmware download operations",
            "type": "apiKey",
            "name": "X-Device-Key",
            "in": "header"
        }
    }
}